Change TimeZone configuration comment and behavior. It is not defined as follow:
        Adjust timezone to use when SquidAnalyzer reports different time
        than graphs in your browser. The value must follow format: +/-HH.
        Default is to use local time. This must be considered as real
        timezone but the number of hours to add remove from log timestamp to
        adjust graphs times. For example:

                TimeZone       +01

        will append one hour to all graphs time series to match squid server
        timestamp.

Be warn that if you have enabled this directive before this commit the
graphs might display the wrong time. If your browser is at the same
timezone as the squid server you might want to comment this directive as
SquidAnalyzer will adjust the graphs to the right timezone by default.

Thanks to Weverton Everaldo Lubask for the report.